AbstractChoiceField.php + RadioButtons.php + other related changes
