+++ src/Form/RadioButtons.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MyParcelCom\JsonSchema\FormBuilder\Form;
+
+class RadioButtons extends AbstractChoiceField
+{
+    protected ChoiceFieldType $fieldType = ChoiceFieldType::RADIO;
+}
