+++ src/Form/Select.php
@@ -4,34 +4,7 @@ declare(strict_types=1);
 
 namespace MyParcelCom\JsonSchema\FormBuilder\Form;
 
-use InvalidArgumentException;
-use MyParcelCom\JsonSchema\FormBuilder\Properties\JsonSchemaProperty;
-use MyParcelCom\JsonSchema\FormBuilder\Properties\PropertyType;
-
-class Select implements Field
+class Select extends AbstractChoiceField
 {
-    public function __construct(
-        public readonly string $name,
-        public readonly PropertyType $type,
-        public readonly string $label,
-        public readonly bool $isRequired = false,
-        public readonly array $enum = [],
-        public readonly ?string $help = null,
-    ) {
-        if (count($enum) < 1) {
-            throw new InvalidArgumentException('Select property requires at least one enum value.');
-        }
-    }
-
-    public function toJsonSchemaProperty(): JsonSchemaProperty
-    {
-        return new JsonSchemaProperty(
-            name: $this->name,
-            type: $this->type,
-            description: $this->label,
-            isRequired: $this->isRequired,
-            enum: $this->enum,
-            help: $this->help,
-        );
-    }
+    protected ChoiceFieldType $fieldType = ChoiceFieldType::SELECT;
 }
